@@ -17,12 +17,12 @@ class ProductsServiceTest extends TestCase
         parent::setUp();
         Model::unguard();
 
-        $this->gas = factory(\App\Gas::class)->create();
-        $this->supplier = factory(\App\Supplier::class)->create();
-        $this->category = factory(\App\Category::class)->create();
-        $this->measure = factory(\App\Measure::class)->create();
+        $this->gas = \App\Gas::factory()->create();
+        $this->supplier = \App\Supplier::factory()->create();
+        $this->category = \App\Category::factory()->create();
+        $this->measure = \App\Measure::factory()->create();
 
-        $this->product = factory(\App\Product::class)->create([
+        $this->product = \App\Product::factory()->create([
             'supplier_id' => $this->supplier->id,
             'category_id' => $this->category->id,
             'measure_id' => $this->measure->id
@@ -30,7 +30,7 @@ class ProductsServiceTest extends TestCase
 
         $this->userWithAdminPerm = $this->createRoleAndUser($this->gas, 'supplier.add');
         $this->userWithReferrerPerms = $this->createRoleAndUser($this->gas, 'supplier.modify', $this->supplier);
-        $this->userWithNoPerms = factory(\App\User::class)->create(['gas_id' => $this->gas->id]);
+        $this->userWithNoPerms = \App\User::factory()->create(['gas_id' => $this->gas->id]);
 
         Model::reguard();
 
