@@ -5,6 +5,7 @@ namespace App;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 use Auth;
 use URL;
@@ -15,7 +16,7 @@ use App\Events\AttachableToGas;
 
 class Aggregate extends Model
 {
-    use GASModel, ModifiableTrait, ReducibleTrait;
+    use HasFactory, GASModel, ModifiableTrait, ReducibleTrait;
 
     protected $dispatchesEvents = [
         'created' => AttachableToGas::class
