@@ -3,6 +3,7 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 use App\Events\SluggableCreating;
 use App\GASModel;
@@ -11,7 +12,7 @@ use App\Hierarchic;
 
 class Category extends Model implements Hierarchic
 {
-    use GASModel, SluggableID;
+    use HasFactory, GASModel, SluggableID;
 
     public $incrementing = false;
     protected $keyType = 'string';
