@@ -3,6 +3,7 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 use App\Events\SluggableCreating;
 use App\GASModel;
@@ -10,7 +11,7 @@ use App\SluggableID;
 
 class Measure extends Model
 {
-    use GASModel, SluggableID;
+    use HasFactory, GASModel, SluggableID;
 
     public $incrementing = false;
     protected $keyType = 'string';
