@@ -3,6 +3,7 @@
 namespace App\Providers;
 
 use Illuminate\Support\ServiceProvider;
+
 use Auth;
 use Menu;
 
@@ -17,55 +18,47 @@ class MenuServiceProvider extends ServiceProvider
                 $user = Auth::user();
                 $gas = $user->gas;
 
-                $menu = Menu::handler('main');
+                Menu::make('MainMenu', function($menu) use ($user, $gas) {
+                    $menu->add('<span class="glyphicon glyphicon-home" aria-hidden="true"></span> ' . _i('Home'), 'dashboard');
+                    $menu->add('<span class="glyphicon glyphicon-cog" aria-hidden="true"></span> ' . $user->printableName(), 'users/profile');
 
-                $menu->add('dashboard', '<span class="glyphicon glyphicon-home" aria-hidden="true"></span> ' . _i('Home'));
-                $menu->add('users/profile', '<span class="glyphicon glyphicon-cog" aria-hidden="true"></span> ' . $user->printableName());
+                    if (is_null($user->suspended_at)) {
+                        if ($user->can('users.admin', $gas) || $user->can('users.view', $gas)) {
+                            $menu->add('<span class="glyphicon glyphicon-user" aria-hidden="true"></span> ' . _i('Utenti'), 'users');
+                        }
 
-                if (is_null($user->suspended_at)) {
-                    if ($user->can('users.admin', $gas) || $user->can('users.view', $gas)) {
-                        $menu->add('users', '<span class="glyphicon glyphicon-user" aria-hidden="true"></span> ' . _i('Utenti'));
-                    }
+                        if ($user->can('supplier.view', $gas) || $user->can('supplier.add', $gas) || $user->can('supplier.modify', null)) {
+                            $menu->add('<span class="glyphicon glyphicon-tags" aria-hidden="true"></span> ' . _i('Fornitori'), 'suppliers');
+                        }
 
-                    if ($user->can('supplier.view', $gas) || $user->can('supplier.add', $gas) || $user->can('supplier.modify', null)) {
-                        $menu->add('suppliers', '<span class="glyphicon glyphicon-tags" aria-hidden="true"></span> ' . _i('Fornitori'));
-                    }
+                        if ($user->can('supplier.orders', null) || $user->can('supplier.shippings', null) || $user->can('order.view', $gas)) {
+                            $menu->add('<span class="glyphicon glyphicon-list-alt" aria-hidden="true"></span> ' . _i('Ordini'), 'orders');
+                        }
 
-                    if ($user->can('supplier.orders', null) || $user->can('supplier.shippings', null) || $user->can('order.view', $gas)) {
-                        $menu->add('orders', '<span class="glyphicon glyphicon-list-alt" aria-hidden="true"></span> ' . _i('Ordini'));
-                    }
+                        if ($user->can('supplier.book', null)) {
+                            $menu->add('<span class="glyphicon glyphicon-bookmark" aria-hidden="true"></span> ' . _i('Prenotazioni'), 'bookings');
+                        }
 
-                    if ($user->can('supplier.book', null)) {
-                        $menu->add('bookings', '<span class="glyphicon glyphicon-bookmark" aria-hidden="true"></span> ' . _i('Prenotazioni'));
-                    }
+                        if ($user->can('movements.view', $gas) || $user->can('movements.admin', $gas)) {
+                            $menu->add('<span class="glyphicon glyphicon-euro" aria-hidden="true"></span> ' . _i('Contabilità'), 'movements');
+                        }
 
-                    if ($user->can('movements.view', $gas) || $user->can('movements.admin', $gas)) {
-                        $menu->add('movements', '<span class="glyphicon glyphicon-euro" aria-hidden="true"></span> ' . _i('Contabilità'));
-                    }
+                        if ($user->can('gas.statistics', $gas)) {
+                            $menu->add('<span class="glyphicon glyphicon-stats" aria-hidden="true"></span> ' . _i('Statistiche'), 'stats');
+                        }
 
-                    if ($user->can('gas.statistics', $gas)) {
-                        $menu->add('stats', '<span class="glyphicon glyphicon-stats" aria-hidden="true"></span> ' . _i('Statistiche'));
-                    }
+                        $menu->add('<span class="glyphicon glyphicon-ok-circle" aria-hidden="true"></span> ' . _i('Notifiche'), 'notifications');
 
-                    $menu->add('notifications', '<span class="glyphicon glyphicon-ok-circle" aria-hidden="true"></span> ' . _i('Notifiche'));
+                        if ($user->can('gas.config', $gas)) {
+                            $menu->add('<span class="glyphicon glyphicon-wrench" aria-hidden="true"></span> ' . _i('Configurazioni'), 'gas/' . $gas->id . '/edit');
+                        }
 
-                    if ($user->can('gas.config', $gas)) {
-                        $menu->add('gas/'.$gas->id.'/edit', '<span class="glyphicon glyphicon-wrench" aria-hidden="true"></span> ' . _i('Configurazioni'));
-                    }
-
-                    if ($user->can('gas.multi', $gas)) {
-                        $menu->add('multigas', '<span class="glyphicon glyphicon-globe" aria-hidden="true"></span> ' . _i('Multi-GAS'));
-                    }
-                }
-
-                $menu->addClass('nav navbar-nav')->getItemsByContentType(Menu\Items\Contents\Link::class)->map(function ($item) {
-                    if ($item->isActive()) {
-                        $item->addClass('active');
+                        if ($user->can('gas.multi', $gas)) {
+                            $menu->add('<span class="glyphicon glyphicon-globe" aria-hidden="true"></span> ' . _i('Multi-GAS'), 'multigas');
+                        }
                     }
                 });
             }
-
-            $view->with('menu', $menu);
         });
     }
 
