@@ -44,20 +44,22 @@
         <nav class="navbar navbar-default navbar-fixed-top navbar-inverse">
             <div class="container-fluid">
                 <div class="navbar-header">
-                    <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
+                    <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#main-navbar" aria-expanded="false">
                         <span class="sr-only">Toggle navigation</span>
                         <span class="icon-bar"></span>
                         <span class="icon-bar"></span>
                         <span class="icon-bar"></span>
                     </button>
-                    @if(!isset($menu))
+                    @if(!isset($MainMenu))
                         <a class="navbar-brand hidden-md hidden-sm" href="{{ route('dashboard') }}">GASdotto</a>
                     @endif
                 </div>
 
-                <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
-                    @if(isset($menu))
-                        {!! $menu !!}
+                <div class="collapse navbar-collapse" id="main-navbar">
+                    @if(isset($MainMenu))
+                        <ul class="nav navbar-nav">
+                            @include(config('laravel-menu.views.bootstrap-items'), ['items' => $MainMenu->roots()])
+                        </ul>
                     @endif
 
                     @if(Auth::check())
